Agregando validacion S/N - Metodo Elimnar Productos-categorias

+++ InvertAppParallel/Category.cs
@@ -32,7 +32,7 @@ namespace InvertAppParallel
             Console.WriteLine("LISTADO DE CATEGORIAS");
             Console.WriteLine("\n");
             ListarCategoria();
-            Console.WriteLine("Introduce el id de una categoria para editar...");
+            Console.WriteLine("\n\nIntroduce el id de una categoria para editar...");
             int idCategory = int.Parse(Console.ReadLine());
 
             Console.WriteLine("Introduce el nombre de la categoria");
@@ -63,15 +63,38 @@ namespace InvertAppParallel
 
         public void EliminarCategoria()
         {
-
-            Console.WriteLine("Introduce el id de una categoria para eliminar...");
+            Console.WriteLine("LISTADO DE CATEGORIAS");
+            Console.WriteLine("\n");
+            ListarCategoria();
+            Console.WriteLine("\n\nIntroduce el id de una categoria para eliminar...");
             int idCategoryEliminar = int.Parse(Console.ReadLine());
 
             using (var db = new InverAppHomeworkDBContext())
             {
-                var data1 = db.Categories.First(a => a.CategoryId == idCategoryEliminar);
-                db.Categories.Remove(data1);
-                db.SaveChanges();
+                Console.WriteLine("Estas seguro que quieres eliminar esta categoria Y/N?");
+                string yesNo = Console.ReadLine();
+
+                if (yesNo == "Y" || yesNo == "y")
+                {
+                    var data1 = db.Categories.First(a => a.CategoryId == idCategoryEliminar);
+                    db.Categories.Remove(data1);
+                    db.SaveChanges();
+                }
+                else if (yesNo == "N" || yesNo == "n")
+                {
+                    Console.Clear();
+                    Menu.mantenimientoCategorias();
+                }
+                else
+                {
+                    Console.Clear();
+                    Console.WriteLine("Has introducido una tecla incorrecta. Vuelve a intentarlo." +
+                                       "Presiona cualquier tecla para salir...");
+                    Menu.mantenimientoCategorias();
+                    Console.ReadKey();
+                }
+
+                
 
             }
         }
